feat(ui): Agent UI v2 Phase 4c - retire server-rendered HTML pages
The v2 SPA (/agent/app/) now covers every server-rendered page: Home /
Agents / Config map to the Agents view, Schedules to the agent detail,
and Chat to the SPA Chat. Retire the HTML pages, keeping the data plane
(every /api/*, /ui/events SSE, serveSpa) fully intact.

Page routes (GET /agents, /jobs, /home, /admin, /, and the EXACT /ui)
now 302 to the SPA with a RELATIVE Location (app/ or app/chat) so the
redirect resolves both daemon-direct and hub-proxied without the daemon
knowing its public mount. Operator bookmarks keep working.

The /ui retirement matches EXACTLY (=== "/ui"), not a prefix, so the
/ui/events message SSE the SPA Chat subscribes to (owned by the http-ui
transport's ingestHttp) keeps routing. /terminal stays a live 200 page,
demoted off the primary nav (an attach tool); ui-kit.ts stays for it.

Deleted: src/home-ui.ts, src/agents-ui.ts (+ test), src/admin-ui.ts
(+ test), src/jobs-ui.ts (+ test), and the inline CHAT_UI_HTML page
block in daemon.ts; removed the now-dead imports (HOME_UI_HTML,
AGENTS_UI_HTML, JOBS_UI_HTML, renderAdminPage, THEME_CSS/appShell/
SHELL_JS). The admin-ui.test.ts module.json contract block moved to
src/module-manifest.test.ts (preserved coverage).

module.json: uiUrl + configUiUrl now point at /agent/app/ (the retired
/agent/home and /agent/admin pages are gone).

Tests: redirect tests for each retired route; a footgun guard proving
GET /ui/events still routes to the SSE handler (401, not a 302); a
/terminal-still-200 test; a representative /api/* still-gated test.
