+++ .parachute/module.json
@@ -7,8 +7,8 @@
   "paths": ["/agent"],
   "health": "/health",
   "stripPrefix": true,
-  "uiUrl": "/agent/home",
-  "configUiUrl": "/agent/admin",
+  "uiUrl": "/agent/app/",
+  "configUiUrl": "/agent/app/",
   "websocket": true,
   "focus": "experimental",
   "adminCapabilities": ["config"],
